+++ react/CheckoutB2B.tsx
@@ -1,3 +1,4 @@
+import { QueryClientProvider } from '@tanstack/react-query'
 import React, { useCallback } from 'react'
 import { useIntl } from 'react-intl'
 import { useCssHandles } from 'vtex.css-handles'
@@ -11,6 +12,7 @@ import {
   Table,
   ToastProvider,
   Totalizer,
+  withToast,
 } from 'vtex.styleguide'
 import './styles.css'
 
@@ -22,11 +24,13 @@ import {
   useToolbar,
   useTotalizers,
 } from './hooks'
+import { queryClient } from './services'
+import { WithToast } from './typings'
 import { messages } from './utils'
 
 const { useOrderItems } = OrderItems
 
-function CheckoutB2B() {
+function CheckoutB2B({ showToast }: WithToast) {
   const handles = useCssHandles(['container', 'table'])
   const { organization, loading: organizationLoading } = useOrganization()
   const {
@@ -39,7 +43,7 @@ function CheckoutB2B() {
   const { items } = orderForm
   const totalizers = useTotalizers()
   const schema = useTableSchema()
-  const toolbar = useToolbar()
+  const toolbar = useToolbar(showToast)
   const { navigate } = useRuntime()
   const { formatMessage } = useIntl()
   const { removeItem } = useOrderItems()
@@ -53,48 +57,58 @@ function CheckoutB2B() {
   console.log('ORDER FORM:', orderForm)
 
   return (
-    <ToastProvider positioning="window">
-      <div className={handles.container}>
-        <Layout
-          fullWidth
-          pageHeader={
-            <PageHeader
-              title={<ExtensionPoint id="rich-text" />}
-              linkLabel={formatMessage(messages.backToHome)}
-              onLinkClick={() => navigate({ page: 'store.home' })}
-            />
-          }
-        >
-          <PageBlock>
-            {!loading && (
-              <div className="mb4">
-                <ContactInfos organization={organization} />
-                <Totalizer items={totalizers} />
-              </div>
-            )}
-
-            <div className={handles.table}>
-              <Table
-                loading={loading}
-                fullWidth
-                schema={schema}
-                items={items}
-                density="high"
-                emptyStateLabel={formatMessage(messages.emptyCart)}
-                toolbar={toolbar}
-              />
+    <div className={handles.container}>
+      <Layout
+        fullWidth
+        pageHeader={
+          <PageHeader
+            title={<ExtensionPoint id="rich-text" />}
+            linkLabel={formatMessage(messages.backToHome)}
+            onLinkClick={() => navigate({ page: 'store.home' })}
+          />
+        }
+      >
+        <PageBlock>
+          {!loading && (
+            <div className="mb4">
+              <ContactInfos organization={organization} />
+              <Totalizer items={totalizers} />
             </div>
-          </PageBlock>
-
-          {!!items.length && !loading && (
-            <Button variation="danger-tertiary" onClick={handleClearCart}>
-              {formatMessage(messages.clearCart)}
-            </Button>
           )}
-        </Layout>
-      </div>
-    </ToastProvider>
+
+          <div className={handles.table}>
+            <Table
+              loading={loading}
+              fullWidth
+              schema={schema}
+              items={items}
+              density="high"
+              emptyStateLabel={formatMessage(messages.emptyCart)}
+              toolbar={toolbar}
+            />
+          </div>
+        </PageBlock>
+
+        {!!items.length && !loading && (
+          <Button variation="danger-tertiary" onClick={handleClearCart}>
+            {formatMessage(messages.clearCart)}
+          </Button>
+        )}
+      </Layout>
+    </div>
   )
 }
 
-export default CheckoutB2B
+const CheckoutB2BWithToast = withToast(CheckoutB2B)
+
+function CheckoutB2BWrapper() {
+  return (
+    <QueryClientProvider client={queryClient}>
+      <ToastProvider positioning="window">
+        <CheckoutB2BWithToast />
+      </ToastProvider>
+    </QueryClientProvider>
+  )
+}
+
+export default CheckoutB2BWrapper