+++ react/components/PONumber.tsx
@@ -1,22 +1,43 @@
-import React, { useState } from 'react'
+import React from 'react'
 import { useIntl } from 'react-intl'
 import { Input } from 'vtex.styleguide'
 
-import { messages } from '../utils'
+import { useOrderFormCustom } from '../hooks'
+import {
+  B2B_CUSTOM_APP_ID,
+  B2B_CUSTOM_APP_MAJOR,
+  messages,
+  PO_NUMBER_CUSTOM_FIELD,
+} from '../utils'
 
 export function PONumber() {
   const { formatMessage } = useIntl()
-  const [selectedPONumber, setSelectedPoNumber] = useState<string>()
+  const { orderForm, setOrderForm } = useOrderFormCustom()
 
   const handleChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    setSelectedPoNumber(e.target.value)
+    setOrderForm({
+      ...orderForm,
+      customData: !e.target.value
+        ? null
+        : {
+            ...orderForm.customData,
+            customApps: [
+              {
+                id: B2B_CUSTOM_APP_ID,
+                major: B2B_CUSTOM_APP_MAJOR,
+                fields: {
+                  [PO_NUMBER_CUSTOM_FIELD]: e.target.value,
+                },
+              },
+            ],
+          },
+    })
   }
 
   return (
     <Input
       size="small"
-      placeholder={formatMessage(messages.PONumber)}
-      value={selectedPONumber}
+      placeholder={formatMessage(messages.PONumberPlaceholder)}
       onChange={handleChange}
     />
   )