+++ react/hooks/useTableSchema.tsx
@@ -1,21 +1,26 @@
 import React from 'react'
 import { useIntl } from 'react-intl'
-import { Item } from 'vtex.checkout-graphql'
+import type { Item } from 'vtex.checkout-graphql'
 import { FormattedPrice } from 'vtex.formatted-price'
-import { useRuntime } from 'vtex.render-runtime'
+import { OrderItems } from 'vtex.order-items'
+import { ButtonWithIcon, IconDelete } from 'vtex.styleguide'
 
 import { QuantitySelector } from '../components/QuantitySelector'
 import { TruncatedText } from '../components/TruncatedText'
-import { TableSchema } from '../typings'
+import type { TableSchema } from '../typings'
 import { isWithoutStock, messages, normalizeString } from '../utils'
+import { useOrderFormCustom } from './useOrderFormCustom'
+
+const { useOrderItems } = OrderItems
 
 function getStrike(item: Item) {
   return { strike: isWithoutStock(item) }
 }
 
 export function useTableSchema(): TableSchema<Item> {
-  const { account } = useRuntime()
+  const { orderForm } = useOrderFormCustom()
   const { formatMessage } = useIntl()
+  const { removeItem } = useOrderItems()
 
   return {
     properties: {
@@ -73,14 +78,13 @@ export function useTableSchema(): TableSchema<Item> {
         width: 150,
         title: formatMessage(messages.seller),
         cellRenderer({ rowData }) {
-          const seller =
-            rowData.seller === '1'
-              ? account.charAt(0).toUpperCase() + account.slice(1)
-              : rowData.seller
-
-          return (
-            <TruncatedText text={seller ?? 'N/A'} {...getStrike(rowData)} />
+          const seller = orderForm.sellers?.find(
+            (s) => rowData.seller === s?.id
           )
+
+          const sellerName = seller?.name ?? rowData.seller ?? 'N/A'
+
+          return <TruncatedText text={sellerName} {...getStrike(rowData)} />
         },
       },
       sellingPrice: {
@@ -117,6 +121,22 @@ export function useTableSchema(): TableSchema<Item> {
                 {...getStrike(rowData)}
               />
             )
+          )
+        },
+      },
+      id: {
+        width: 50,
+        title: ' ',
+        cellRenderer({ rowData }) {
+          return (
+            <ButtonWithIcon
+              size="small"
+              icon={<IconDelete />}
+              variation="danger-tertiary"
+              onClick={() => {
+                removeItem({ id: rowData.id, seller: rowData.seller ?? '1' })
+              }}
+            />
           )
         },
       },