+++ react/hooks/useToolbar.ts
@@ -1,10 +1,13 @@
 import { useIntl } from 'react-intl'
 
+import type { WithToast } from '../typings'
 import { messages } from '../utils'
 import { useOrderFormCustom } from './useOrderFormCustom'
+import { usePlaceOrder } from './usePlaceOrder'
 
-export function useToolbar() {
+export function useToolbar(showToast: WithToast['showToast']) {
   const { formatMessage } = useIntl()
+  const { placeOrder, isLoading, isSuccess } = usePlaceOrder(showToast)
   const { orderForm } = useOrderFormCustom()
 
   if (!orderForm?.items?.length) return null
@@ -21,11 +24,10 @@ export function useToolbar() {
       onSubmit: console.log,
     },
     newLine: {
-      disabled: false,
-      isLoading: false,
+      disabled: isLoading || isSuccess,
+      isLoading,
       label: formatMessage(messages.placeOrder),
-      // eslint-disable-next-line no-console
-      handleCallback: () => console.log(formatMessage(messages.placeOrder)),
+      handleCallback: placeOrder,
     },
   }
 }