+++ react/utils/index.ts
@@ -1,6 +1,10 @@
-import { Item } from 'vtex.checkout-graphql'
+import { Item, PaymentData } from 'vtex.checkout-graphql'
 
 export * from './messages'
+
+export const B2B_CUSTOM_APP_ID = 'b2b-checkout-settings'
+export const B2B_CUSTOM_APP_MAJOR = 1
+export const PO_NUMBER_CUSTOM_FIELD = 'purchaseOrderNumber'
 
 export function normalizeString(str?: string | null) {
   return (
@@ -13,4 +17,15 @@ export function normalizeString(str?: string | null) {
 
 export function isWithoutStock(item: Item) {
   return item.availability === 'withoutStock'
+}
+
+export function getFirstInstallmentByPaymentSystem(
+  installmentOptions: PaymentData['installmentOptions'],
+  paymentSystem?: string | null
+) {
+  const installmentOption = installmentOptions.find(
+    (option) => option.paymentSystem === paymentSystem
+  )
+
+  return installmentOption?.installments[0]
 }